Remove timestamp from serialization
The UI reacts badly when we add the timestamp on every span.
Looking at http://zipkin.io/pages/instrumenting.html tells us that we
should only add the timestamp on span originators (i.e. excluding sr/ss
spans if they comes from an instrumented client). However, root span
should send it.

We remove it for the moment since the UI tries to guess the timestamp
and because we really need to refactor the way we handle tracing.

Change-Id: I2091fb84bd58d4ad75278fbd07ab9fa1cb9d6168