Trace is MarshalledByRef instead of Serializable
"When an object derives form MarshalByRefObject,
an object reference will be passed from one application
domain to another rather than the object itself."

It should fix the issue we have been encountering
with tests where there are different app domains.
Object was serialized but the destination app domain
wasn't able to deserialize the object as it couldn't locate the tracing assembly.

Change-Id: Id92d3fd2e1c964f8ea7bc5eba20cf8c7e2a5bd7b